split passes
